refactor(controller): update the algorithm
Use a user factory to retrieve already instantiated User object. I have to do this
because, there are other parts of the source code which is going to require the
same instance of this object. I don't prefer caching this object on the controller,
as the controller is intended to provide behaviors to the architecture