refactor(credit service): refactor setBalance
Made sure that it's smart enough to know if it needs to increment the user
balance OR if it needs to decrement the user balance. I have added the
increase argument to the setBalance function to determine this purpose.

I have also extracted out the real property value I need to represent
the payment argument